refactor(login): improve login page maintainability
