+++ lib/views/navigation/routes.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 import "package:proxima/views/pages/home_page.dart";
-import "package:proxima/views/pages/login_page.dart";
+import "package:proxima/views/pages/login/login_page.dart";
 
 enum Routes {
   home("home"),
