+++ test/component/login_test.dart
@@ -3,7 +3,8 @@ import "package:flutter_test/flutter_test.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/main.dart";
 import "package:proxima/views/pages/home_page.dart";
-import "package:proxima/views/pages/login_page.dart";
+import "package:proxima/views/pages/login/login_button.dart";
+import "package:proxima/views/pages/login/login_page.dart";
 
 import "utils/firebase/setup_firebase_mocks.dart";
 import "utils/firebase/testing_login_providers.dart";
@@ -33,7 +34,7 @@ void main() {
     final sloganFinder = find.text("Discover the world,\n one post at a time");
     expect(sloganFinder, findsOneWidget);
 
-    final loginButton = find.byKey(LoginPage.loginButtonKey);
+    final loginButton = find.byKey(LoginButton.loginButtonKey);
     // Check that the login button is displayed and contains the "Login" text
     expect(
       find.descendant(
@@ -59,7 +60,7 @@ void main() {
     await tester.pumpWidget(mockedProxima);
     await tester.pumpAndSettle();
 
-    final loginButton = find.byKey(LoginPage.loginButtonKey);
+    final loginButton = find.byKey(LoginButton.loginButtonKey);
     await tester.tap(loginButton);
     await tester.pumpAndSettle();
